Make -base compile almost without warnings again.  Fixes:
- Using the correct format specifier
- Remove substringFromRange: (GNUstep extension) from NSString.h and use the non-deprecated version in -base.
- Delete definitions of methods that are the same in both a class and a category on that class (NSArray and NSObject).

Two files still have warnings:

- The GNUTLS stuff is using a load of deprecated APIs.
- xmlparse.m replaces a method using a category.  This method seems to always return nil.  Not sure what it's meant to do, or why it isn't a delegate of some kind.



git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@33756 72102866-910b-0410-8b05-ffd578937521

+++ Tools/gdnc.m
@@ -536,7 +536,7 @@ ihandler(int sig)
 
   if (debugging)
     NSLog(@"Adding observer %llu for %@ %@",
-      anObserver, notificationName, anObject);
+      (unsigned long long)anObserver, notificationName, anObject);
 
   connection = [(NSDistantObject*)client connectionForProxy];
   clients = (NSMapTable*)NSMapGet(connections, connection);
@@ -818,7 +818,7 @@ ihandler(int sig)
 		{
 		  [obs->queue removeObjectAtIndex: 0];
   if (debugging)
-    NSLog(@"Posting to observer %llu with %@", obs->observer, n);
+    NSLog(@"Posting to observer %llu with %@", (unsigned long long)obs->observer, n);
 		  [obs->client->client postNotificationName: n->name
 						     object: n->object
 						   userInfo: n->info
@@ -842,7 +842,7 @@ ihandler(int sig)
 {
   if (debugging)
     NSLog(@"Removing observer %llu for %@ %@",
-      observer->observer, observer->notificationName,
+      (unsigned long long)observer->observer, observer->notificationName,
       observer->notificationObject);
 
   if (observer->notificationObject)
